# packages/contracts/openapi/control-v1.yaml
...
openapi: 3.1.0
info:
  title: OLO ToolGate Control Plane API
  version: 0.10.0-dev
  description: Tenant-scoped group administration with reviewed configuration changes
    and online current authorization. Signed snapshots describe configuration; execution
    requires a separately signed single-use permit.
security:
- adminAccessToken: []
paths:
...
  /api/control/v1/tools/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        $ref: ../schemas/v1/identifiers.schema.json#/$defs/Identifier
    get:
      operationId: getControlTool
      description: Requires toolgate-admin or toolgate-reader.
      responses:
        '200':
          description: Success
          headers:
            ETag:
              description: Quoted revision
              schema:
                type: string
            X-Request-ID:
              description: Server correlation ID
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/control.schema.json#/$defs/ControlTool
        '400': *id002
        '401': *id003
        '403': *id004
        '404': *id005
        '409': *id006
        '413': *id007
        '415': *id008
        '500': *id009
        '503': *id010
    put:
      operationId: updateControlTool
      description: Requires toolgate-admin.
      responses:
        '200':
          description: Success
          headers:
            ETag:
              description: Quoted revision
              schema:
                type: string
            X-Request-ID:
              description: Server correlation ID
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/control.schema.json#/$defs/ControlTool
        '400': *id002
        '401': *id003
        '403': *id004
        '404': *id005
        '409': *id006
        '413': *id007
        '415': *id008
        '500': *id009
        '503': *id010
        '202':
          description: Draft created; review impact, submit for independent review,
            then apply
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseConfigurationChange
      parameters:
      - *id011
      - *id012
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/control.schema.json#/$defs/ControlTool
    delete:
      operationId: deleteControlTool
      description: Requires toolgate-admin.
      responses:
        '204':
          description: Success
          headers:
            ETag:
              description: Quoted revision
              schema:
                type: string
            X-Request-ID:
              description: Server correlation ID
              schema:
                type: string
        '400': *id002
        '401': *id003
        '403': *id004
        '404': *id005
        '409': *id006
        '413': *id007
        '415': *id008
        '500': *id009
        '503': *id010
        '202':
          description: Draft created; review impact, submit for independent review,
            then apply
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseConfigurationChange
      parameters:
      - *id011
      - *id012
...
components:
  securitySchemes:
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 token with verified issuer/audience/expiry and signed tenant_id/groups.
        Not a Gateway runtime credential.
    deviceCertificate:
      type: mutualTLS
      description: Direct verified device TLS identity; forwarded identity headers
        never authenticate.